requirements.txt
streamlit>=1.32.0
langchain>=0.1.12
langchain-groq>=0.1.3
langchain-community>=0.0.28
langchain-openai>=0.1.0
langchain-core>=0.1.30
requests>=2.31.0
pandas>=2.2.1
plotly>=5.19.0
web3>=6.15.1
python-dotenv>=1.0.1
apscheduler>=3.10.4
scikit-learn>=1.4.1.post1
numpy>=1.26.4
dune-client>=1.4.0
flipside>=2.1.0
sqlalchemy>=2.0.28
aiohttp>=3.9.3
asyncio>=3.4.3
